requirements.txt
######### python=3.7, torch=1.10
absl-py==1.0.0
appdirs==1.4.4
audioread==2.1.9
boto3==1.20.21
botocore==1.23.21
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
decorator==5.1.0
faiss-gpu==1.7.1.post3
filelock==3.4.0
ftfy==6.0.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
graphviz==0.8.4
grpcio==1.42.0
huggingface-hub==0.1.2
idna==3.3
importlib-metadata==4.8.2
jmespath==0.10.0
joblib==1.1.0
librosa==0.6.0
llvmlite==0.31.0
Markdown==3.3.6
mxnet==1.8.0.post0
nltk==3.7
numba==0.48.0
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.4.60
packaging==21.3
Pillow==8.4.0
pip==21.2.2
pooch==1.5.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.6
python-dateutil==2.8.2
pytorch-pretrained-bert==0.6.2
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.2.1
setuptools==58.0.4
six==1.16.0
SoundFile==0.10.3.post1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
threadpoolctl==3.0.0
timm==0.4.12
tokenizers==0.10.3
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3
transformers==4.11.3
typing_extensions==4.0.0
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wheel==0.37.0
yacs==0.1.8
zhon==1.1.5
zipp==3.6.0
